Shield et Health system avec degat au joueur qui update bien UI #9

+++ dev/FasterThanPixels/FasterThanPixels/src/ECS/Components/C_Health.h
@@ -7,5 +7,7 @@
 */
 class C_Health : public Engine::BlankComponent {
 public:
-	float Health;  /*!< Vie de l'entite */
+	int Health;  /*!< Vie de l'entite */
 };
+
+//
